pyproject.toml
[project]
name = "cronix"
version = "0.1.0"
description = "A lightweight, production-ready task scheduling system based on FastAPI and PostgreSQL"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Cronix Team" }
]
keywords = ["fastapi", "scheduler", "cron", "task", "postgresql"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "fastapi[standard]==0.128.0",
    "sqlalchemy[asyncio]==2.0.45",
    "asyncpg==0.30.0",
    "pydantic-settings==2.1.0",
    "croniter==6.0.0",
    "aiohttp==3.11.11",
    "pyjwt==2.10.1",
    "bcrypt==5.0.0",
    "pyotp==2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
]

[project.urls]
Homepage = "https://github.com/oxzk/cronix"
Repository = "https://github.com/oxzk/cronix"
Issues = "https://github.com/oxzk/cronix/issues"

[project.scripts]
cronix = "main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by formatter
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
